chore(model): RPC conversion code isolation (E-I)
