+++ crates/api-model/src/instance/config/network.rs
@@ -19,12 +19,10 @@ use std::collections::{HashMap, HashSet};
 use std::fmt::Display;
 use std::net::IpAddr;
 
-use ::rpc::errors::RpcDataConversionError;
 use carbide_uuid::machine::MachineId;
 use carbide_uuid::network::{NetworkPrefixId, NetworkSegmentId};
 use carbide_uuid::vpc::VpcPrefixId;
 use ipnetwork::IpNetwork;
-use itertools::Itertools;
 use mac_address::MacAddress;
 use serde::ser::SerializeMap;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
@@ -37,26 +35,6 @@ use crate::ConfigValidationError;
 pub enum InterfaceFunctionType {
     Physical = 0,
     Virtual = 1,
-}
-
-impl TryFrom<rpc::InterfaceFunctionType> for InterfaceFunctionType {
-    type Error = RpcDataConversionError;
-
-    fn try_from(function_type: rpc::InterfaceFunctionType) -> Result<Self, Self::Error> {
-        Ok(match function_type {
-            rpc::InterfaceFunctionType::Physical => InterfaceFunctionType::Physical,
-            rpc::InterfaceFunctionType::Virtual => InterfaceFunctionType::Virtual,
-        })
-    }
-}
-
-impl From<InterfaceFunctionType> for rpc::InterfaceFunctionType {
-    fn from(function_type: InterfaceFunctionType) -> rpc::InterfaceFunctionType {
-        match function_type {
-            InterfaceFunctionType::Physical => rpc::InterfaceFunctionType::Physical,
-            InterfaceFunctionType::Virtual => rpc::InterfaceFunctionType::Virtual,
-        }
-    }
 }
 
 /// Uniquely identifies an interface on the instance
@@ -397,299 +375,6 @@ impl InstanceNetworkConfig {
     }
 }
 
-#[derive(PartialEq)]
-enum VFAllocationType {
-    // Only physical interface is defined. No virtual function is defined.
-    None,
-    // Cloud is sending valid virtual function id.
-    Cloud,
-    // Cloud is sending None for virtual function id. This bis possible in older versions.
-    Carbide,
-}
-
-type DeviceVFIdsMap =
-    HashMap<(Option<String>, u32), Vec<(rpc::InterfaceFunctionType, Option<u32>)>>;
-
-fn validate_virtual_function_ids_and_get_allocation_method(
-    interfaces: &[rpc::InstanceInterfaceConfig],
-) -> Result<VFAllocationType, RpcDataConversionError> {
-    let mut device_vf_ids: DeviceVFIdsMap = HashMap::new();
-
-    // Create grouping based on device and device_instance.
-    interfaces.iter().for_each(|x| {
-        device_vf_ids
-            .entry((x.device.clone(), x.device_instance))
-            .or_default()
-            .push((x.function_type(), x.virtual_function_id))
-    });
-
-    let all_vf_ids = device_vf_ids
-        .values()
-        .flatten()
-        .filter(|x| x.0 == rpc::InterfaceFunctionType::Virtual)
-        .collect_vec();
-
-    if all_vf_ids.is_empty() {
-        // Only Physical interfaces are mentioned.
-        return Ok(VFAllocationType::None);
-    }
-
-    if all_vf_ids.iter().all(|x| x.1.is_none()) {
-        // Virtual function ids are not yet implemented at cloud.
-        return Ok(VFAllocationType::Carbide);
-    }
-
-    if all_vf_ids.iter().any(|x| x.1.is_none()) {
-        // At least one None and one valid virtual_function_id is given. Mix of both is not allowed.
-        return Err(RpcDataConversionError::InvalidValue(
-            "Mix of VF".to_string(),
-            "Mix of valid virtual_function_id and None is found.".to_string(),
-        ));
-    }
-
-    for vf_info in device_vf_ids.values() {
-        let vf_ids = vf_info
-            .iter()
-            .filter_map(|(ft, vf_id)| {
-                if let rpc::InterfaceFunctionType::Virtual = ft {
-                    Some(*vf_id)
-                } else {
-                    None
-                }
-            })
-            .flatten()
-            .collect_vec();
-
-        if vf_ids.is_empty() {
-            // Only physical interfaces are provided.
-            // Nothing to validate for this device and device_instance.
-            continue;
-        }
-
-        // Check for duplicate VF ids.
-        let vf_ids_set = vf_ids.iter().collect::<HashSet<&u32>>();
-        if vf_ids.len() != vf_ids_set.len() {
-            return Err(RpcDataConversionError::InvalidValue(
-                "Duplicate VFs".to_string(),
-                "Duplicate VF IDs detected.".to_string(),
-            ));
-        }
-    }
-
-    // All device and device_instance's VF IDs are validated.
-    Ok(VFAllocationType::Cloud)
-}
-
-impl TryFrom<rpc::InstanceNetworkConfig> for InstanceNetworkConfig {
-    type Error = RpcDataConversionError;
-
-    fn try_from(config: rpc::InstanceNetworkConfig) -> Result<Self, Self::Error> {
-        // try_from for interfaces:
-        let auto = config.auto;
-
-        if auto && !config.interfaces.is_empty() {
-            return Err(RpcDataConversionError::InvalidArgument(
-                "InstanceNetworkConfig.auto cannot be combined with explicit interfaces"
-                    .to_string(),
-            ));
-        }
-
-        let mut assigned_vfs_map: HashMap<(Option<String>, u32), u8> = HashMap::default();
-        let mut interfaces = Vec::with_capacity(config.interfaces.len());
-        // Either all virtual ids for VF are None, or all should have some valid values.
-        // virtual_function_id can not be repeated.
-
-        let allocation_type =
-            validate_virtual_function_ids_and_get_allocation_method(&config.interfaces)?;
-        for iface in config.interfaces.into_iter() {
-            let rpc_iface_type = rpc::InterfaceFunctionType::try_from(iface.function_type)
-                .map_err(|_| {
-                    RpcDataConversionError::InvalidInterfaceFunctionType(iface.function_type)
-                })?;
-            let iface_type = InterfaceFunctionType::try_from(rpc_iface_type).map_err(|_| {
-                RpcDataConversionError::InvalidInterfaceFunctionType(iface.function_type)
-            })?;
-
-            let function_id = match iface_type {
-                InterfaceFunctionType::Physical => InterfaceFunctionId::Physical {},
-                InterfaceFunctionType::Virtual => {
-                    // Note that this might overflow if the RPC call delivers more than
-                    // 256 VFs. However that's ok - the `InstanceNetworkConfig.validate()`
-                    // call will declare those configs as invalid later on anyway.
-                    // We mainly don't want to crash here.
-                    InterfaceFunctionId::Virtual {
-                        id: if allocation_type == VFAllocationType::Carbide {
-                            let assigned_vfs = assigned_vfs_map
-                                .entry((iface.device.clone(), iface.device_instance))
-                                .or_insert(0);
-                            let id = *assigned_vfs;
-                            *assigned_vfs = assigned_vfs.saturating_add(1);
-                            id
-                        } else {
-                            // Already validated.
-                            iface.virtual_function_id.unwrap_or_default() as u8
-                        },
-                    }
-                }
-            };
-
-            // If network_details is present, that gets precedence and we'll pull the network_segment_id from that
-            // if it's a NetworkSegment.
-            let (network_details, network_segment_id) = if let Some(x) = iface.network_details {
-                let nd: NetworkDetails = x.try_into()?;
-                let ns_id = match nd {
-                    NetworkDetails::NetworkSegment(network_segment_id) => Some(network_segment_id),
-                    NetworkDetails::VpcPrefixId(_uuid) => None,
-                };
-
-                (Some(nd), ns_id)
-            } else {
-                // If network_details wasn't set, then the caller is required to
-                // send network_segment_id.
-                // This is old model. Let's use network segment id as such.
-                // TODO: This should be removed in future.
-                let ns_id =
-                    iface
-                        .network_segment_id
-                        .ok_or(RpcDataConversionError::MissingArgument(
-                            "InstanceInterfaceConfig::network_segment_id",
-                        ))?;
-
-                // And then we'll populate network_details from that as well.
-                (Some(NetworkDetails::NetworkSegment(ns_id)), Some(ns_id))
-            };
-
-            if iface.ip_address.is_some()
-                && matches!(network_details, Some(NetworkDetails::NetworkSegment(..)))
-            {
-                return Err(RpcDataConversionError::InvalidArgument(
-                    "explicit IP requests are only supported for VPC prefixes".to_string(),
-                ));
-            };
-
-            // ipv6_interface_config is only valid alongside a VPC prefix -- it makes no
-            // sense with a NetworkSegment (segment already has its own prefixes) or
-            // without any network_details at all. Check before parsing.
-            if iface.ipv6_interface_config.is_some()
-                && !matches!(network_details, Some(NetworkDetails::VpcPrefixId(_)))
-            {
-                return Err(RpcDataConversionError::InvalidArgument(
-                    "ipv6 requires vpc_prefix_id to be set".to_string(),
-                ));
-            };
-
-            // Prevent setting an IPv6 address in ip_address when ipv6_interface_config
-            // is also set -- that would mean two IPv6 configs for the same interface,
-            // and DHCP can't hand out two IPs of the same family on one interface.
-            if let Some(ref ip_str) = iface.ip_address
-                && iface.ipv6_interface_config.is_some()
-                && ip_str.parse::<std::net::Ipv6Addr>().is_ok()
-            {
-                return Err(RpcDataConversionError::InvalidArgument(
-                    "ip_address cannot be IPv6 when ipv6_interface_config is also set -- use ipv6_interface_config.ip_address for the IPv6 address".to_string(),
-                ));
-            }
-
-            let ipv6_interface_config = iface
-                .ipv6_interface_config
-                .map(
-                    |v6| -> Result<Ipv6InterfaceConfig, RpcDataConversionError> {
-                        let vpc_prefix_id =
-                            v6.vpc_prefix_id
-                                .ok_or(RpcDataConversionError::MissingArgument(
-                                    "InstanceInterfaceIpv6Config::vpc_prefix_id",
-                                ))?;
-                        let requested_ip_addr = v6
-                            .ip_address
-                            .map(|s| {
-                                s.parse::<std::net::Ipv6Addr>().map_err(|_| {
-                                    RpcDataConversionError::InvalidIpAddress(format!(
-                                        "IPv6 address expected, got: {s}"
-                                    ))
-                                })
-                            })
-                            .transpose()?;
-                        Ok(Ipv6InterfaceConfig {
-                            vpc_prefix_id,
-                            requested_ip_addr,
-                        })
-                    },
-                )
-                .transpose()?;
-
-            let device_locator = iface.device.map(|device| DeviceLocator {
-                device,
-                device_instance: iface.device_instance as usize,
-            });
-
-            interfaces.push(InstanceInterfaceConfig {
-                function_id,
-                network_segment_id,
-                network_details,
-                ip_addrs: HashMap::default(),
-                requested_ip_addr: iface
-                    .ip_address
-                    .map(|i| i.parse::<IpAddr>())
-                    .transpose()
-                    .map_err(|e| RpcDataConversionError::InvalidIpAddress(e.to_string()))?,
-                ipv6_interface_config,
-                interface_prefixes: HashMap::default(),
-                network_segment_gateways: HashMap::new(),
-                host_inband_mac_address: None,
-                device_locator,
-                internal_uuid: uuid::Uuid::new_v4(),
-            });
-        }
-
-        Ok(Self { interfaces, auto })
-    }
-}
-
-impl TryFrom<InstanceNetworkConfig> for rpc::InstanceNetworkConfig {
-    type Error = RpcDataConversionError;
-
-    fn try_from(config: InstanceNetworkConfig) -> Result<rpc::InstanceNetworkConfig, Self::Error> {
-        let auto = config.auto;
-        let mut interfaces = Vec::with_capacity(config.interfaces.len());
-        for iface in config.interfaces.into_iter() {
-            let function_type = iface.function_id.function_type();
-
-            // Update network segment id based on network details.
-            let network_details: Option<rpc::forge::instance_interface_config::NetworkDetails> =
-                iface.network_details.map(|x| x.into());
-            let network_segment_id = iface.network_segment_id;
-
-            let (device, device_instance) = match iface.device_locator {
-                Some(dl) => (Some(dl.device), dl.device_instance as u32),
-                None => (None, 0),
-            };
-
-            let virtual_function_id = match iface.function_id {
-                InterfaceFunctionId::Physical {} => None,
-                InterfaceFunctionId::Virtual { id } => Some(id as u32),
-            };
-
-            interfaces.push(rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::from(function_type) as i32,
-                network_segment_id,
-                network_details,
-                device,
-                device_instance,
-                virtual_function_id,
-                ip_address: iface.requested_ip_addr.map(|i| i.to_string()),
-                ipv6_interface_config: iface.ipv6_interface_config.map(|v6| {
-                    rpc::forge::InstanceInterfaceIpv6Config {
-                        vpc_prefix_id: Some(v6.vpc_prefix_id),
-                        ip_address: v6.requested_ip_addr.map(|i| i.to_string()),
-                    }
-                }),
-            });
-        }
-
-        Ok(rpc::InstanceNetworkConfig { interfaces, auto })
-    }
-}
-
 /// Validates that any container which has elements that have InterfaceFunctionIds
 /// assigned assigned is using unique and valid FunctionIds.
 pub fn validate_interface_function_ids<
@@ -775,35 +460,6 @@ pub enum NetworkDetails {
     VpcPrefixId(VpcPrefixId),
 }
 
-impl From<NetworkDetails> for rpc::forge::instance_interface_config::NetworkDetails {
-    fn from(value: NetworkDetails) -> Self {
-        match value {
-            NetworkDetails::NetworkSegment(network_segment_id) => {
-                rpc::forge::instance_interface_config::NetworkDetails::SegmentId(network_segment_id)
-            }
-            NetworkDetails::VpcPrefixId(uuid) => {
-                rpc::forge::instance_interface_config::NetworkDetails::VpcPrefixId(uuid)
-            }
-        }
-    }
-}
-
-impl TryFrom<rpc::forge::instance_interface_config::NetworkDetails> for NetworkDetails {
-    fn try_from(
-        value: rpc::forge::instance_interface_config::NetworkDetails,
-    ) -> Result<Self, Self::Error> {
-        Ok(match value {
-            rpc::forge::instance_interface_config::NetworkDetails::SegmentId(ns_id) => {
-                NetworkDetails::NetworkSegment(ns_id)
-            }
-            rpc::forge::instance_interface_config::NetworkDetails::VpcPrefixId(vpc_prefix_id) => {
-                NetworkDetails::VpcPrefixId(vpc_prefix_id)
-            }
-        })
-    }
-
-    type Error = RpcDataConversionError;
-}
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize, Hash, Default)]
 pub struct DeviceLocator {
     pub device: String,
@@ -1080,106 +736,6 @@ mod tests {
     }
 
     #[test]
-    fn assign_ids_from_rpc_config_pf_only() {
-        let config = rpc::InstanceNetworkConfig {
-            interfaces: vec![rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Physical as _,
-                network_segment_id: Some(NetworkSegmentId::from(BASE_SEGMENT_ID)),
-                network_details: None,
-                device: None,
-                device_instance: 0u32,
-                virtual_function_id: None,
-                ip_address: None,
-                ipv6_interface_config: None,
-            }],
-            auto: false,
-        };
-
-        let netconfig: InstanceNetworkConfig = config.try_into().unwrap();
-        assert_eq!(
-            netconfig.interfaces,
-            &[InstanceInterfaceConfig {
-                function_id: InterfaceFunctionId::Physical {},
-                network_segment_id: Some(BASE_SEGMENT_ID.into()),
-                ip_addrs: HashMap::new(),
-                requested_ip_addr: None,
-                ipv6_interface_config: None,
-                interface_prefixes: HashMap::new(),
-                network_segment_gateways: HashMap::new(),
-                host_inband_mac_address: None,
-                network_details: Some(NetworkDetails::NetworkSegment(BASE_SEGMENT_ID.into()),),
-                device_locator: None,
-                internal_uuid: netconfig.interfaces.first().unwrap().internal_uuid,
-            }]
-        );
-    }
-
-    #[test]
-    fn assign_ids_from_rpc_config_pf_and_vf() {
-        let mut interfaces = vec![rpc::InstanceInterfaceConfig {
-            function_type: rpc::InterfaceFunctionType::Physical as _,
-            network_segment_id: Some(BASE_SEGMENT_ID.into()),
-            network_details: None,
-            device: None,
-            device_instance: 0u32,
-            virtual_function_id: None,
-            ip_address: None,
-            ipv6_interface_config: None,
-        }];
-        for vfid in INTERFACE_VFID_MIN..=INTERFACE_VFID_MAX {
-            interfaces.push(rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Virtual as _,
-                network_segment_id: Some(offset_segment_id(vfid + 1)),
-                network_details: None,
-                device: None,
-                device_instance: 0u32,
-                virtual_function_id: None,
-                ip_address: None,
-                ipv6_interface_config: None,
-            });
-        }
-
-        let config = rpc::InstanceNetworkConfig {
-            interfaces,
-            auto: false,
-        };
-        let netconfig: InstanceNetworkConfig = config.try_into().unwrap();
-        let mut netconf_interfaces_iter = netconfig.interfaces.iter();
-
-        let mut expected_interfaces = vec![InstanceInterfaceConfig {
-            function_id: InterfaceFunctionId::Physical {},
-            network_segment_id: Some(BASE_SEGMENT_ID.into()),
-            ip_addrs: HashMap::new(),
-            requested_ip_addr: None,
-            ipv6_interface_config: None,
-            interface_prefixes: HashMap::new(),
-            network_segment_gateways: HashMap::new(),
-            host_inband_mac_address: None,
-            network_details: Some(NetworkDetails::NetworkSegment(BASE_SEGMENT_ID.into())),
-            device_locator: None,
-            internal_uuid: netconf_interfaces_iter.next().unwrap().internal_uuid,
-        }];
-
-        for vfid in INTERFACE_VFID_MIN..=INTERFACE_VFID_MAX {
-            let segment_id = offset_segment_id(vfid + 1);
-            expected_interfaces.push(InstanceInterfaceConfig {
-                function_id: InterfaceFunctionId::Virtual { id: vfid },
-                network_segment_id: Some(segment_id),
-                ip_addrs: HashMap::new(),
-                requested_ip_addr: None,
-                ipv6_interface_config: None,
-                interface_prefixes: HashMap::new(),
-                network_segment_gateways: HashMap::new(),
-                host_inband_mac_address: None,
-                network_details: Some(NetworkDetails::NetworkSegment(segment_id)),
-                device_locator: None,
-                internal_uuid: netconf_interfaces_iter.next().unwrap().internal_uuid,
-            });
-        }
-        assert_eq!(netconfig.interfaces, &expected_interfaces[..]);
-    }
-
-    #[test]
     fn validate_network_config() {
         let config = create_valid_network_config();
         config.validate(true).unwrap();
@@ -1217,411 +773,5 @@ mod tests {
         config.interfaces[0].network_segment_id = Some(DUPLICATE_SEGMENT_ID.into());
         config.interfaces[1].network_segment_id = Some(DUPLICATE_SEGMENT_ID.into());
         assert!(config.validate(true).is_err());
-    }
-
-    fn get_rpc_instance_network_config() -> Vec<rpc::InstanceInterfaceConfig> {
-        vec![
-            rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Physical as i32,
-                network_segment_id: None,
-                virtual_function_id: None,
-                network_details: Some(
-                    rpc::forge::instance_interface_config::NetworkDetails::SegmentId(
-                        offset_segment_id(0),
-                    ),
-                ),
-                device: None,
-                device_instance: 0u32,
-                ip_address: None,
-                ipv6_interface_config: None,
-            },
-            rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Virtual as i32,
-                network_segment_id: None,
-                virtual_function_id: Some(0),
-                network_details: Some(
-                    rpc::forge::instance_interface_config::NetworkDetails::SegmentId(
-                        offset_segment_id(1),
-                    ),
-                ),
-                device: None,
-                device_instance: 0u32,
-                ip_address: None,
-                ipv6_interface_config: None,
-            },
-            rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Virtual as i32,
-                network_segment_id: None,
-                virtual_function_id: Some(1),
-                network_details: Some(
-                    rpc::forge::instance_interface_config::NetworkDetails::SegmentId(
-                        offset_segment_id(2),
-                    ),
-                ),
-                device: None,
-                device_instance: 0u32,
-                ip_address: None,
-                ipv6_interface_config: None,
-            },
-            rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Virtual as i32,
-                network_segment_id: None,
-                virtual_function_id: Some(2),
-                network_details: Some(
-                    rpc::forge::instance_interface_config::NetworkDetails::SegmentId(
-                        offset_segment_id(3),
-                    ),
-                ),
-                device: None,
-                device_instance: 0u32,
-                ip_address: None,
-                ipv6_interface_config: None,
-            },
-        ]
-    }
-
-    #[test]
-    fn test_validate_virtual_function_ids() {
-        let interfaces = get_rpc_instance_network_config();
-
-        let network_config = rpc::InstanceNetworkConfig {
-            interfaces,
-            auto: false,
-        };
-        let network_config: InstanceNetworkConfig = network_config.try_into().unwrap();
-
-        let vf_ids = network_config.interfaces.iter().filter_map(|x| {
-            if let InterfaceFunctionId::Virtual { id } = x.function_id {
-                Some(id)
-            } else {
-                None
-            }
-        });
-
-        let vf_ids = vf_ids.sorted().collect_vec();
-
-        // All VF ids should be present after converting.
-        let expected = vec![0, 1, 2];
-        assert_eq!(expected, vf_ids);
-    }
-
-    #[test]
-    fn test_validate_virtual_function_ids_missing_1() {
-        let mut interfaces = get_rpc_instance_network_config();
-        interfaces.remove(2);
-
-        let network_config = rpc::InstanceNetworkConfig {
-            interfaces,
-            auto: false,
-        };
-        let network_config: InstanceNetworkConfig = network_config.try_into().unwrap();
-
-        let vf_ids = network_config.interfaces.iter().filter_map(|x| {
-            if let InterfaceFunctionId::Virtual { id } = x.function_id {
-                Some(id)
-            } else {
-                None
-            }
-        });
-
-        let vf_ids = vf_ids.sorted().collect_vec();
-
-        // Since vf_id: 1 is removed, it should not be present in the parsed config.
-        let expected = vec![0, 2];
-        assert_eq!(expected, vf_ids);
-    }
-
-    #[test]
-    fn test_validate_virtual_function_ids_only_physical() {
-        let mut interfaces = get_rpc_instance_network_config();
-        interfaces = vec![interfaces[0].clone()];
-
-        let network_config = rpc::InstanceNetworkConfig {
-            interfaces,
-            auto: false,
-        };
-        let network_config: InstanceNetworkConfig = network_config.try_into().unwrap();
-
-        let vf_ids = network_config
-            .interfaces
-            .iter()
-            .filter_map(|x| {
-                if let InterfaceFunctionId::Virtual { id } = x.function_id {
-                    Some(id)
-                } else {
-                    None
-                }
-            })
-            .collect_vec();
-
-        assert!(vf_ids.is_empty());
-    }
-
-    #[test]
-    fn test_validate_virtual_function_ids_duplicate() {
-        let mut interfaces = get_rpc_instance_network_config();
-        interfaces[2].virtual_function_id = Some(0);
-
-        let network_config = rpc::InstanceNetworkConfig {
-            interfaces,
-            auto: false,
-        };
-        let network_config: Result<InstanceNetworkConfig, RpcDataConversionError> =
-            network_config.try_into();
-
-        assert!(network_config.is_err());
-    }
-
-    #[test]
-    fn test_validate_virtual_function_ids_mix() {
-        let mut interfaces = get_rpc_instance_network_config();
-        interfaces[2].virtual_function_id = None;
-
-        let network_config = rpc::InstanceNetworkConfig {
-            interfaces,
-            auto: false,
-        };
-        let network_config: Result<InstanceNetworkConfig, RpcDataConversionError> =
-            network_config.try_into();
-
-        assert!(network_config.is_err());
-    }
-
-    #[test]
-    fn test_network_details_serde_backward_compat_single() {
-        // Old JSON format: single VPC prefix.
-        let uuid = uuid::Uuid::new_v4();
-        let json = format!(r#"{{"VpcPrefixId":"{}"}}"#, uuid);
-        let nd: NetworkDetails = serde_json::from_str(&json).unwrap();
-        assert_eq!(nd, NetworkDetails::VpcPrefixId(VpcPrefixId::from(uuid)));
-
-        // Round-trip
-        let serialized = serde_json::to_string(&nd).unwrap();
-        let nd2: NetworkDetails = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(nd, nd2);
-    }
-
-    #[test]
-    fn test_network_details_rpc_roundtrip_single() {
-        let id = VpcPrefixId::new();
-        let model_nd = NetworkDetails::VpcPrefixId(id);
-
-        // Model -> RPC
-        let rpc_nd: rpc::forge::instance_interface_config::NetworkDetails = model_nd.clone().into();
-        assert!(matches!(
-            rpc_nd,
-            rpc::forge::instance_interface_config::NetworkDetails::VpcPrefixId(_)
-        ));
-
-        // RPC -> Model
-        let roundtripped: NetworkDetails = rpc_nd.try_into().unwrap();
-        assert_eq!(roundtripped, model_nd);
-    }
-
-    #[test]
-    fn test_dual_stack_rpc_roundtrip() {
-        // Verify that ipv6 survives a model → rpc → model round-trip.
-        let v4_id = VpcPrefixId::new();
-        let v6_id = VpcPrefixId::new();
-
-        let model_config = InstanceNetworkConfig {
-            interfaces: vec![InstanceInterfaceConfig {
-                function_id: InterfaceFunctionId::Physical {},
-                network_segment_id: None,
-                network_details: Some(NetworkDetails::VpcPrefixId(v4_id)),
-                ip_addrs: HashMap::default(),
-                requested_ip_addr: None,
-                ipv6_interface_config: Some(Ipv6InterfaceConfig {
-                    vpc_prefix_id: v6_id,
-                    requested_ip_addr: Some("2001:db8::1".parse().unwrap()),
-                }),
-                interface_prefixes: HashMap::default(),
-                network_segment_gateways: HashMap::default(),
-                host_inband_mac_address: None,
-                device_locator: None,
-                internal_uuid: uuid::Uuid::new_v4(),
-            }],
-            auto: false,
-        };
-
-        // Model -> RPC
-        let rpc_config: rpc::InstanceNetworkConfig = model_config.try_into().unwrap();
-        let rpc_iface = &rpc_config.interfaces[0];
-        assert!(matches!(
-            rpc_iface.network_details,
-            Some(rpc::forge::instance_interface_config::NetworkDetails::VpcPrefixId(_))
-        ));
-        assert_eq!(
-            rpc_iface
-                .ipv6_interface_config
-                .as_ref()
-                .and_then(|v6| v6.vpc_prefix_id),
-            Some(v6_id)
-        );
-        assert_eq!(
-            rpc_iface
-                .ipv6_interface_config
-                .as_ref()
-                .and_then(|v6| v6.ip_address.clone()),
-            Some("2001:db8::1".to_string())
-        );
-
-        // RPC -> Model
-        let roundtripped: InstanceNetworkConfig = rpc_config.try_into().unwrap();
-        let v6 = roundtripped.interfaces[0]
-            .ipv6_interface_config
-            .as_ref()
-            .unwrap();
-        assert_eq!(v6.vpc_prefix_id, v6_id);
-        assert_eq!(v6.requested_ip_addr, Some("2001:db8::1".parse().unwrap()));
-    }
-
-    #[test]
-    fn test_ipv6_requires_vpc_prefix_id() {
-        // ipv6 without vpc_prefix_id should be rejected.
-        let v6_id = VpcPrefixId::new();
-        let rpc_config = rpc::InstanceNetworkConfig {
-            interfaces: vec![rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Physical as i32,
-                network_segment_id: Some(NetworkSegmentId::new()),
-                network_details: Some(
-                    rpc::forge::instance_interface_config::NetworkDetails::SegmentId(
-                        NetworkSegmentId::new(),
-                    ),
-                ),
-                device: None,
-                device_instance: 0,
-                virtual_function_id: None,
-                ip_address: None,
-                ipv6_interface_config: Some(rpc::forge::InstanceInterfaceIpv6Config {
-                    vpc_prefix_id: Some(v6_id),
-                    ip_address: None,
-                }),
-            }],
-            auto: false,
-        };
-        let result: Result<InstanceNetworkConfig, _> = rpc_config.try_into();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_v6_only_uses_primary_field() {
-        // v6-only allocation: just put the v6 prefix in the primary vpc_prefix_id field.
-        let v6_id = VpcPrefixId::new();
-        let rpc_config = rpc::InstanceNetworkConfig {
-            interfaces: vec![rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Physical as i32,
-                network_segment_id: None,
-                network_details: Some(
-                    rpc::forge::instance_interface_config::NetworkDetails::VpcPrefixId(v6_id),
-                ),
-                device: None,
-                device_instance: 0,
-                virtual_function_id: None,
-                ip_address: None,
-                ipv6_interface_config: None,
-            }],
-            auto: false,
-        };
-        let model: InstanceNetworkConfig = rpc_config.try_into().unwrap();
-        assert_eq!(
-            model.interfaces[0].network_details,
-            Some(NetworkDetails::VpcPrefixId(v6_id))
-        );
-        assert_eq!(model.interfaces[0].ipv6_interface_config, None);
-    }
-
-    #[test]
-    fn test_reject_v6_ip_address_with_ipv6_interface_config() {
-        // Setting an IPv6 ip_address AND ipv6_interface_config should be rejected.
-        let v4_id = VpcPrefixId::new();
-        let v6_id = VpcPrefixId::new();
-        let rpc_config = rpc::InstanceNetworkConfig {
-            interfaces: vec![rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Physical as i32,
-                network_segment_id: None,
-                network_details: Some(
-                    rpc::forge::instance_interface_config::NetworkDetails::VpcPrefixId(v4_id),
-                ),
-                device: None,
-                device_instance: 0,
-                virtual_function_id: None,
-                ip_address: Some("2001:db8::1".to_string()),
-                ipv6_interface_config: Some(rpc::forge::InstanceInterfaceIpv6Config {
-                    vpc_prefix_id: Some(v6_id),
-                    ip_address: None,
-                }),
-            }],
-            auto: false,
-        };
-        let result: Result<InstanceNetworkConfig, _> = rpc_config.try_into();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_allow_v4_ip_address_with_ipv6_interface_config() {
-        // Setting an IPv4 ip_address AND ipv6_interface_config is fine (dual-stack).
-        let v4_id = VpcPrefixId::new();
-        let v6_id = VpcPrefixId::new();
-        let rpc_config = rpc::InstanceNetworkConfig {
-            interfaces: vec![rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Physical as i32,
-                network_segment_id: None,
-                network_details: Some(
-                    rpc::forge::instance_interface_config::NetworkDetails::VpcPrefixId(v4_id),
-                ),
-                device: None,
-                device_instance: 0,
-                virtual_function_id: None,
-                ip_address: Some("10.0.0.1".to_string()),
-                ipv6_interface_config: Some(rpc::forge::InstanceInterfaceIpv6Config {
-                    vpc_prefix_id: Some(v6_id),
-                    ip_address: None,
-                }),
-            }],
-            auto: false,
-        };
-        let result: Result<InstanceNetworkConfig, _> = rpc_config.try_into();
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_auto_rejects_explicit_interfaces() {
-        // `auto: true` and populated interfaces are mutually exclusive,
-        // so this should error.
-        let rpc_config = rpc::InstanceNetworkConfig {
-            interfaces: vec![rpc::InstanceInterfaceConfig {
-                function_type: rpc::InterfaceFunctionType::Physical as i32,
-                network_segment_id: Some(NetworkSegmentId::new()),
-                network_details: None,
-                device: None,
-                device_instance: 0,
-                virtual_function_id: None,
-                ip_address: None,
-                ipv6_interface_config: None,
-            }],
-            auto: true,
-        };
-        let result: Result<InstanceNetworkConfig, _> = rpc_config.try_into();
-        let err = result.expect_err("auto + non-empty interfaces should be rejected");
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("auto"),
-            "error should mention auto, got: {msg}"
-        );
-    }
-
-    #[test]
-    fn test_auto_allows_empty_interfaces() {
-        // Verify "auto" requests work.
-        let rpc_config = rpc::InstanceNetworkConfig {
-            interfaces: vec![],
-            auto: true,
-        };
-        let model: InstanceNetworkConfig = rpc_config
-            .try_into()
-            .expect("auto + empty should round-trip");
-        assert!(model.auto);
-        assert!(model.interfaces.is_empty());
     }
 }
