+++ crates/api-model/src/instance/status/extension_service.rs
@@ -17,12 +17,10 @@
 
 use std::collections::HashMap;
 
-use ::rpc::errors::RpcDataConversionError;
 use carbide_uuid::extension_service::ExtensionServiceId;
 use carbide_uuid::machine::MachineId;
 use chrono::{DateTime, Utc};
 use config_version::{ConfigVersion, Versioned};
-use rpc::forge as rpc;
 use serde::{Deserialize, Serialize};
 
 use crate::extension_service::ExtensionServiceType;
@@ -38,21 +36,6 @@ pub struct InstanceExtensionServicesStatus {
 
     /// Whether all desired extension service changes that the user has applied have taken effect
     pub configs_synced: SyncState,
-}
-
-impl TryFrom<InstanceExtensionServicesStatus> for rpc::InstanceDpuExtensionServicesStatus {
-    type Error = RpcDataConversionError;
-
-    fn try_from(status: InstanceExtensionServicesStatus) -> Result<Self, Self::Error> {
-        let mut extension_services = Vec::with_capacity(status.extension_services.len());
-        for service in status.extension_services {
-            extension_services.push(rpc::InstanceDpuExtensionServiceStatus::try_from(service)?);
-        }
-        Ok(rpc::InstanceDpuExtensionServicesStatus {
-            dpu_extension_services: extension_services,
-            configs_synced: rpc::SyncState::try_from(status.configs_synced)? as i32,
-        })
-    }
 }
 
 impl InstanceExtensionServicesStatus {
@@ -315,108 +298,12 @@ pub enum ExtensionServiceDeploymentStatus {
     Error,
 }
 
-impl From<rpc::DpuExtensionServiceDeploymentStatus> for ExtensionServiceDeploymentStatus {
-    fn from(status: rpc::DpuExtensionServiceDeploymentStatus) -> Self {
-        match status {
-            rpc::DpuExtensionServiceDeploymentStatus::DpuExtensionServiceUnknown => Self::Unknown,
-            rpc::DpuExtensionServiceDeploymentStatus::DpuExtensionServicePending => Self::Pending,
-            rpc::DpuExtensionServiceDeploymentStatus::DpuExtensionServiceRunning => Self::Running,
-            rpc::DpuExtensionServiceDeploymentStatus::DpuExtensionServiceTerminating => {
-                Self::Terminating
-            }
-            rpc::DpuExtensionServiceDeploymentStatus::DpuExtensionServiceTerminated => {
-                Self::Terminated
-            }
-            rpc::DpuExtensionServiceDeploymentStatus::DpuExtensionServiceFailed => Self::Failed,
-            rpc::DpuExtensionServiceDeploymentStatus::DpuExtensionServiceError => Self::Error,
-        }
-    }
-}
-
-impl From<ExtensionServiceDeploymentStatus> for rpc::DpuExtensionServiceDeploymentStatus {
-    fn from(status: ExtensionServiceDeploymentStatus) -> Self {
-        match status {
-            ExtensionServiceDeploymentStatus::Unknown => Self::DpuExtensionServiceUnknown,
-            ExtensionServiceDeploymentStatus::Pending => Self::DpuExtensionServicePending,
-            ExtensionServiceDeploymentStatus::Running => Self::DpuExtensionServiceRunning,
-            ExtensionServiceDeploymentStatus::Terminating => Self::DpuExtensionServiceTerminating,
-            ExtensionServiceDeploymentStatus::Terminated => Self::DpuExtensionServiceTerminated,
-            ExtensionServiceDeploymentStatus::Failed => Self::DpuExtensionServiceFailed,
-            ExtensionServiceDeploymentStatus::Error => Self::DpuExtensionServiceError,
-        }
-    }
-}
-
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ExtensionServiceComponent {
     pub name: String,
     pub version: String, // This is the version of the component, not the version of the extension service
     pub url: String,
     pub status: String,
-}
-
-impl TryFrom<rpc::DpuExtensionServiceComponent> for ExtensionServiceComponent {
-    type Error = RpcDataConversionError;
-
-    fn try_from(component: rpc::DpuExtensionServiceComponent) -> Result<Self, Self::Error> {
-        Ok(Self {
-            name: component.name,
-            version: component.version,
-            url: component.url,
-            status: component.status,
-        })
-    }
-}
-
-impl From<ExtensionServiceComponent> for rpc::DpuExtensionServiceComponent {
-    fn from(component: ExtensionServiceComponent) -> Self {
-        Self {
-            name: component.name,
-            version: component.version,
-            url: component.url,
-            status: component.status,
-        }
-    }
-}
-
-impl TryFrom<MachineExtensionServiceStatus> for rpc::DpuExtensionServiceStatus {
-    type Error = RpcDataConversionError;
-
-    fn try_from(status: MachineExtensionServiceStatus) -> Result<Self, Self::Error> {
-        Ok(Self {
-            dpu_machine_id: Some(status.machine_id),
-            status: rpc::DpuExtensionServiceDeploymentStatus::from(status.status).into(),
-            error_message: status.error_message,
-            components: status
-                .components
-                .into_iter()
-                .map(rpc::DpuExtensionServiceComponent::from)
-                .collect(),
-        })
-    }
-}
-
-impl TryFrom<InstanceExtensionServiceStatus> for rpc::InstanceDpuExtensionServiceStatus {
-    type Error = RpcDataConversionError;
-
-    fn try_from(status: InstanceExtensionServiceStatus) -> Result<Self, Self::Error> {
-        let dpu_statuses = status
-            .dpu_statuses
-            .into_iter()
-            .map(rpc::DpuExtensionServiceStatus::try_from)
-            .collect::<Result<Vec<_>, _>>()?;
-
-        Ok(Self {
-            service_id: status.service_id.into(),
-            version: status.version.to_string(),
-            deployment_status: rpc::DpuExtensionServiceDeploymentStatus::from(
-                status.overall_status,
-            )
-            .into(),
-            dpu_statuses,
-            removed: status.removed,
-        })
-    }
 }
 
 /// A single extension service status reported by DPU agent
@@ -430,87 +317,6 @@ pub struct ExtensionServiceStatusObservation {
     pub overall_state: ExtensionServiceDeploymentStatus,
     pub components: Vec<ExtensionServiceComponent>,
     pub message: String,
-}
-
-impl TryFrom<rpc::DpuExtensionServiceStatusObservation> for ExtensionServiceStatusObservation {
-    type Error = RpcDataConversionError;
-
-    fn try_from(
-        observation: rpc::DpuExtensionServiceStatusObservation,
-    ) -> Result<Self, Self::Error> {
-        let service_id = observation
-            .service_id
-            .parse::<ExtensionServiceId>()
-            .map_err(|e| {
-                RpcDataConversionError::InvalidUuid("ExtensionServiceId", e.to_string())
-            })?;
-
-        let service_type = rpc::DpuExtensionServiceType::try_from(observation.service_type)
-            .map_err(|_| {
-                RpcDataConversionError::InvalidValue(
-                    observation.service_type.to_string(),
-                    "service_type".to_string(),
-                )
-            })?
-            .into();
-
-        let overall_state = rpc::DpuExtensionServiceDeploymentStatus::try_from(observation.state)
-            .map_err(|_| {
-                RpcDataConversionError::InvalidValue(
-                    observation.state.to_string(),
-                    "state".to_string(),
-                )
-            })?
-            .into();
-
-        let components = observation
-            .components
-            .into_iter()
-            .map(ExtensionServiceComponent::try_from)
-            .collect::<Result<Vec<_>, _>>()?;
-
-        let version = observation.version.parse::<ConfigVersion>().map_err(|e| {
-            RpcDataConversionError::InvalidConfigVersion(format!(
-                "Failed to parse version as ConfigVersion: {}",
-                e
-            ))
-        })?;
-
-        Ok(Self {
-            service_id,
-            service_type,
-            service_name: observation.service_name,
-            version,
-            removed: observation.removed,
-            overall_state,
-            components,
-            message: observation.message,
-        })
-    }
-}
-
-impl From<ExtensionServiceStatusObservation> for rpc::DpuExtensionServiceStatusObservation {
-    fn from(observation: ExtensionServiceStatusObservation) -> Self {
-        Self {
-            service_id: observation.service_id.into(),
-            service_type: rpc::DpuExtensionServiceType::from(observation.service_type).into(),
-            service_name: observation.service_name,
-            version: observation.version.to_string(),
-            removed: observation.removed,
-            state: rpc::DpuExtensionServiceDeploymentStatus::from(observation.overall_state).into(),
-            components: observation
-                .components
-                .into_iter()
-                .map(|c| rpc::DpuExtensionServiceComponent {
-                    name: c.name,
-                    version: c.version,
-                    url: c.url,
-                    status: c.status,
-                })
-                .collect(),
-            message: observation.message,
-        }
-    }
 }
 
 /// Observation of extension service statuses reported by a single DPU
